requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas_ta>=0.3.14b
python-dotenv>=1.0.0
PyYAML>=6.0.0
tenacity>=8.2.3
matplotlib>=3.0.0
Flask>=2.0.0
